requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy>=2.0.36
aiosqlite==0.19.0
greenlet>=3.0.0
asyncpg>=0.29.0  # PostgreSQL async driver for production

# Authentication
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.6

# Templating
jinja2==3.1.3

# Configuration
pydantic-settings==2.1.0
email-validator>=2.0.0

# Development
python-dotenv==1.0.0

